refactor(kotest): move generator into generator package, emit into convert
Group the Kotest generator core (KotestWirespecGenerator, KotestField,
KotestOverrides, KotestBuilderJvm, KotestWirespecKotlinGenerator) under a
dedicated generator package, and rename the emit package to convert.
