+++ src/integration/kotest/src/commonMain/kotlin/community/flock/wirespec/integration/kotest/generator/KotestField.kt
@@ -1,4 +1,4 @@
-package community.flock.wirespec.integration.kotest
+package community.flock.wirespec.integration.kotest.generator
 
 import kotlin.reflect.KType
