+++ src/integration/kotest/src/commonTest/kotlin/community/flock/wirespec/integration/kotest/KotestOverrideRegistryTest.kt
@@ -1,5 +1,10 @@
 package community.flock.wirespec.integration.kotest
 
+import community.flock.wirespec.integration.kotest.generator.FieldKey
+import community.flock.wirespec.integration.kotest.generator.IdentityRefinedWrapper
+import community.flock.wirespec.integration.kotest.generator.KotestFieldString
+import community.flock.wirespec.integration.kotest.generator.OverrideRegistry
+import community.flock.wirespec.integration.kotest.generator.PathPattern
 import io.kotest.property.Arb
 import io.kotest.property.arbitrary.constant
 import kotlin.test.Test
